zedi/requirements.txt
Django==2.1.5
djangorestframework
markdown
django-filter
psycopg2-binary
django-extensions
kiteconnect
django-model-utils